Specialize Ring operations with Long instead of Int.
Ring has a morphism from Long and CompleteField has a morphism from
Double. This leads to an ambiguity because Int values can be coerced to
either Long or Double values; this is why I initially went with an Int
morphism.

Long specialization is more generally useful so I'm switching before
its too late (and I sure don't want both). I may lift the Long
specialization to scalar multiplication of vectors, we'll see.

+++ src/main/scala/basis/algebra/CompleteField.scala
@@ -9,7 +9,7 @@ package basis.algebra
 
 /** An element of a complete field.
   * 
-  * $IntMorphismInfo
+  * $LongMorphismInfo
   * 
   * $DoubleMorphismInfo
   * 
@@ -76,18 +76,18 @@ trait CompleteField[CompleteField] extends Field[CompleteField] {
     * $NotAlgebraicallyClosedWarning */
   def sqrt: CompleteField
   
-  def + (n: Int): CompleteField = this + n.toDouble
+  def + (n: Long): CompleteField = this + n.toDouble
   
-  def - (n: Int): CompleteField = this - n.toDouble
+  def - (n: Long): CompleteField = this - n.toDouble
   
-  def * (n: Int): CompleteField = this * n.toDouble
+  def * (n: Long): CompleteField = this * n.toDouble
   
-  def / (n: Int): CompleteField = this / n.toDouble
+  def / (n: Long): CompleteField = this / n.toDouble
   
-  /** Raises this $element to an `Int` power.
+  /** Raises this $element to an `Long` power.
     * 
-    * @param  n   the `Int` exponent.
-    * @return the exponentiation of this $element by the `Int` exponent.
+    * @param  n   the `Long` exponent.
+    * @return the exponentiation of this $element by the `Long` exponent.
     */
-  def pow(n: Int): CompleteField = pow(n.toDouble)
+  def pow(n: Long): CompleteField = pow(n.toDouble)
 }
